refactor: top menu user account & notification dropdown into components
