+++ frontend/app/components/nav-menu/user-account-dropdown.hbs
@@ -1,0 +1,29 @@
+ <li class="nav-item dropdown user-profile-dropdown" ...attributes>
+    <a href="#" class="nav-link dropdown-toggle btn-trigger mr-2" id="user-acc-dd" data-toggle="dropdown"
+       aria-haspopup="true" aria-expanded="false">
+      {{#if this.me.user.profileUri}}
+        <img src={{this.me.user.profileUri}} class="rounded-circle user-profile-image" alt={{t
+                "profile.profile_picture"}}>
+        <span class="pl-2">{{this.me.name}}</span>
+      {{else}}
+        <span aria-hidden="true">&times;</span>
+      {{/if}}
+    </a>
+    <div class="dropdown-menu dropdown-menu-lg-right dropdown-secondary"
+         aria-labelledby="user-acc-dd">
+
+      <LinkTo @route="profile" class="dropdown-item" data-toggle="collapse" data-target="#mainTopNav.show">
+        {{t "labels.user_profile_link" user=this.me.name}}
+      </LinkTo>
+
+      <LinkTo @route="profile" @query={{hash view="Settings"}}  class="dropdown-item"
+              data-toggle="collapse" data-target="#mainTopNav.show">
+        {{t "buttons.settings"}}
+      </LinkTo>
+
+      <a class="dropdown-item" href="#" {{on "click" @logout }}
+         data-toggle="collapse" data-target="#mainTopNav.show">
+        {{t "buttons.logout"}}
+      </a>
+    </div>
+  </li>
