+++ frontend/app/components/nav-menu/user-account-dropdown.js
@@ -1,0 +1,6 @@
+import Component from '@glimmer/component';
+import { inject as service } from '@ember/service';
+
+export default class NavMenuUserAccountDropdownComponent extends Component {
+  @service me;
+}
