+++ frontend/app/styles/_notifications-dropdown.scss
@@ -1,0 +1,184 @@
+$palette: (
+  gray30: #3F484D,
+  gray40: #B3BAC566,
+  primary:#F57010,
+  red: #EA0606,
+  white: #FFFF,
+  blue: #0052CCFF,
+);
+
+.notifications-dropdown {
+  .dd-trigger-button {
+    position: relative;
+    outline: none;
+    box-shadow: none;
+    &:hover{
+      color:  map-get($palette, white) !important;
+    }
+  }
+
+  .dropdown-toggle::after {
+    display: none;
+  }
+
+  .dd-trigger-button svg path {
+    fill: map-get($palette, white);
+  }
+
+  .dd-trigger-button .count {
+    position: absolute;
+    background-color: map-get($palette, red);
+    border-radius: 100px;
+    right: 4px;
+    display: block;
+    text-align: center;
+    top: 7px;
+    font-size: 10px;
+    font-weight: 800;
+    width: 15px;
+    height: 15px;
+  }
+
+  .dropdown-menu.show {
+    .header-title {
+      font-size: 16px;
+    }
+
+    .notification-items {
+      overflow-y: auto;
+      max-height: 440px;
+
+      .notification-item {
+        border-bottom: solid 1px map-get($palette, gray40);
+        padding-right: 8px;
+
+        .notification-container {
+          font-weight: 300;
+          white-space: normal;
+          margin-right: -2px;
+        }
+
+        &.active,
+        &:active {
+          color: inherit;
+          background-color: inherit;
+          outline: none;
+        }
+      }
+    }
+    .empty-state{
+      font-weight: 300;
+    }
+
+  }
+
+  .notification-item {
+    .notification-container {
+      .title {
+        font-size: 14px;
+        font-weight: 400;
+        cursor: pointer;
+        padding-right: 0;
+      }
+
+      .body {
+        padding-left: 1.5rem;
+        font-size: 14px;
+        padding-right: 0;
+      }
+
+      .time {
+        padding-left: 1.5rem;
+        font-size: small;
+        color: #b3bac5;
+        font-weight: 600;
+        padding-right: 0;
+      }
+    }
+
+    .notification-actions {
+      .btn-delete {
+        display: flex;
+        opacity: 0;
+        padding: 0;
+        margin: 0 4px;
+        height: 21px;
+        width: 21px;
+        align-items: center;
+        justify-content: center;
+        border: none;
+        outline: none;
+        background-color: transparent;
+
+        &:hover {
+          background-color: #ebecf0;
+        }
+      }
+
+      .btn-toggle-read-status {
+        padding: 0;
+        height: 16px;
+        width: 16px;
+        border: solid 4px transparent;
+        background-color: transparent;
+        outline: none;
+        border-radius: 50%;
+        margin: 3px 0 0;
+      }
+    }
+
+    &:hover {
+      .notification-actions {
+        .btn-delete {
+          opacity: 1;
+        }
+
+        .btn-toggle-read-status {
+          border-color: #ebecf0;
+          background-color: #ebecf0;
+        }
+      }
+    }
+  }
+
+  .notification-item.unread {
+    font-weight: 600;
+
+    .notification-actions {
+      .btn-toggle-read-status {
+        border: solid 4px #ebecf0;
+        background-color: map-get($palette, blue);
+      }
+    }
+  }
+}
+
+/**
+On large screen
+ */
+@media (min-width: 992px) {
+
+  .notifications-dropdown {
+    .dropdown-menu.show {
+      width: 320px;
+      right: 0 !important;
+      left: auto !important;
+      transform: none !important;
+      top: 100% !important;
+    }
+  }
+  .notifications-dropdown .dropdown-menu.show::after {
+    border-color: transparent transparent map-get($palette, white);
+    border-image: none;
+    border-style: solid;
+    border-width: 8px;
+    bottom: 100%;
+    left: 88%;
+    content: " ";
+    height: 0;
+    pointer-events: none;
+    position: absolute;
+    width: 0;
+    margin-left: 0;
+  }
+}
