+++ frontend/app/styles/_user-acc-dropdown.scss
@@ -1,0 +1,76 @@
+$palette: (
+  white: #FFFF,
+  primary: #F57010
+);
+
+.user-profile-dropdown {
+  .user-profile-image {
+    height: 30px;
+  }
+
+  .btn-trigger:after {
+    position: relative;
+    -webkit-transform: rotate(45deg);
+    -ms-transform: rotate(45deg);
+    transform: rotate(45deg);
+    top: 0;
+    content: "";
+    border: 0.13em solid map-get($palette,white);
+    -webkit-transition: all .3s ease;
+    -o-transition: all .3s ease;
+    transition: all .3s ease;
+    overflow: hidden;
+    border-left: 1px solid transparent;
+    border-top: 1px solid transparent;
+    display: inline-block;
+    width: 8px;
+    height: 8px;
+    margin-left: 9px;
+    opacity: 1;
+  }
+
+}
+
+.signin-menu-link,
+.signup-menu-link {
+  font-size: 15px;
+  letter-spacing: 0;
+  border: 1px solid map-get($palette,white);
+  border-radius: 2px;
+  padding-top: 8px;
+  padding-bottom: 8px;
+  display: inline;
+}
+
+.signup-menu-link {
+  background-color: map-get($palette, white);
+  color: map-get($palette, primary) !important;
+}
+
+.signin-menu-link{
+  color: map-get($palette, white);
+}
+
+/**
+On large screen
+ */
+@media (min-width: 992px) {
+
+
+
+  .user-profile-dropdown .dropdown-menu.show::after {
+    border-color: transparent transparent map-get($palette,white);
+    border-image: none;
+    border-style: solid;
+    border-width: 8px;
+    bottom: 100%;
+    content: " ";
+    height: 0;
+    pointer-events: none;
+    position: absolute;
+    width: 0;
+    left: 55%;
+    margin-left: 0;
+  }
+
+}
