+++ frontend/app/styles/app.scss
@@ -157,3 +157,6 @@ pre {
 @import "rating-and-review";
 @import "teacher/teacher";
 @import "h5p-player";
+@import "notifications-dropdown";
+@import "user-acc-dropdown";
+
